提交服务器代码

+++ Server/Src/LogicServer/PlayerManager.cpp
@@ -29,7 +29,7 @@ BOOL CPlayerManager::ReleasePlayer( UINT64 u64RoleID )
 	return Delete(u64RoleID);
 }
 
-BOOL CPlayerManager::DeletePlayer(UINT32 u64RoleID)
+BOOL CPlayerManager::DeletePlayer(UINT64 u64RoleID)
 {
 	return TRUE;
 }
